pyproject.toml
[tool.poetry]
name = "user-research"
version = "0.1.0"
description = ""
authors = ["Oz"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]
    
[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = "^0.83.0"
langchain = "^0.3.9"
python-dotenv = "^1.0.1"
anthropic = "^0.40.0"
openai = "^1.56.0"
pillow = "^11.0.0"
streamlit = "^1.40.2"
plotly = "^5.24.1"
pandas = "^2.2.3"
networkx = "^3.4.2"
groq = "^0.13.0"
langchain-groq = "^0.2.1"
langchain-ollama = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
